+++ packages/components/psammead-copyright/src/index.stories.jsx
@@ -1,9 +1,10 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { withReadme } from 'storybook-readme';
-import Readme from '../README.md';
+import notes from '../README.md';
 import Copyright from './index';
 
-storiesOf('Copyright', module)
-  .addDecorator(withReadme(Readme))
-  .add('default', () => <Copyright>Getty Images</Copyright>);
+storiesOf('Copyright', module).add(
+  'default',
+  () => <Copyright>Getty Images</Copyright>,
+  { notes },
+);