+++ packages/components/psammead-figure/src/index.stories.jsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { withReadme } from 'storybook-readme';
 import Image from '@bbc/psammead-image';
-import Readme from '../README.md';
+import notes from '../README.md';
 import Figure from '.';
 
 const imageAlt =
@@ -10,10 +9,12 @@ const imageAlt =
 const imageSrc =
   'https://ichef.bbci.co.uk/news/640/cpsprodpb/439A/production/_100960371_syrians_and_asylum_v2-nc.png';
 const imageWidth = 853;
-storiesOf('Figure', module)
-  .addDecorator(withReadme(Readme))
-  .add('containing Image', () => (
+storiesOf('Figure', module).add(
+  'containing Image',
+  () => (
     <Figure>
       <Image alt={imageAlt} src={imageSrc} width={imageWidth} />
     </Figure>
-  ));
+  ),
+  { notes },
+);